Seperated SourceBreakdown component
